requirements.txt
discord.py
Pillow
asyncio